% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CB_Method_2024.R
\name{CB_Method}
\alias{CB_Method}
\title{Conservative Balance (CB)}
\usage{
CB_Method(data, Means = F, seed = 123456L)
}
\arguments{
\item{data}{Data frame containing sediment source and one mixture}

\item{Means}{Boolean to switch when using mean and sd data}

\item{seed}{Seed for the random number generator}
}
\value{
Data frame transformed to scalar variable for further analysis.
}
\description{
The function read the isotopic ratio and content data of each individual tracer in your dataset and combine them together with the mixture to create scalar variables while preserving the same exact information.
}
