% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LDA_Plot.R
\name{LDAPlot}
\alias{LDAPlot}
\title{Linear discriminant analysis chart}
\usage{
LDAPlot(data, P3D = FALSE, text = FALSE, colors = NULL, interactive = FALSE)
}
\arguments{
\item{data}{Data frame containing source and mixtures data}

\item{P3D}{Boolean to switch between 2 to 3 dimensional chart}

\item{text}{Boolean to show or not the identification number of each sample point in the plot}

\item{colors}{Allows choosing between a different set of colors in the plots}

\item{interactive}{Boolean to determine whether the plot should be interactive}
}
\description{
The function performs a linear discriminant analysis and displays the data in the relevant dimensions.
}
