% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ternary_diagram.R,
%   R/Ternary_diagram_cluster_new.R
\name{Ternary_diagram}
\alias{Ternary_diagram}
\title{Ternary diagrams with the CI function results}
\usage{
Ternary_diagram(
  data,
  tracers = 1:2,
  n_row = 1,
  n_col = 2,
  Means = TRUE,
  cluster = TRUE,
  dendro_lvls = 5,
  vrtl_sol = FALSE,
  vrtl = c(0.05, 0.9, 0.05)
)

Ternary_diagram(
  data,
  tracers = 1:2,
  n_row = 1,
  n_col = 2,
  Means = TRUE,
  cluster = TRUE,
  dendro_lvls = 5,
  vrtl_sol = FALSE,
  vrtl = c(0.05, 0.9, 0.05)
)
}
\arguments{
\item{data}{Data frame containing the results from the CI_Method function}

\item{tracers}{Number of tracers to display}

\item{n_row}{Number of rows in which to display the tracers}

\item{n_col}{Number of columns in which to display the tracers}

\item{Means}{Boolean indicating whether to use mean and SD data (TRUE) or raw data (FALSE)}

\item{cluster}{Boolean indicating whether to order the triangles based on a clustering function}

\item{dendro_lvls}{Number of branch levels in the clustering dendrogram}

\item{vrtl_sol}{Boolean indicating whether to plot a theoretical solution}

\item{vrtl}{Numeric vector of values for the theoretical solution}
}
\value{
Ternary diagrams from the individual tracer predictions

Ternary diagrams from the individual tracer predictions
}
\description{
Display the CI prediction for the selected number

Display the CI prediction for the selected number
}
