% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unmix_R_legacy.R
\name{unmix_R_legacy}
\alias{unmix_R_legacy}
\title{Unmix sediment mixtures}
\usage{
unmix_R_legacy(data, iter = "short", Means = F, VA = "SEM", seed = 123456L)
}
\arguments{
\item{data}{Data frame containing sediment source and mixtures}

\item{iter}{Iterations in the source variability analysis}

\item{Means}{Boolean to switch when using mean and sd data}

\item{VA}{Boolean to switch between analyzing the SEM or the SD of the sources}

\item{seed}{Seed for the random number generator}
}
\value{
Data frame containing the relative contribution of the sediment sources for each sediment mixture and iterations
}
\description{
Asses the relative contribution of the potential sediment sources for each sediment mixture in the dataset.
}
